Move the files in the `reports` directory into the `pages/reports`.
