+++ js/src/pages/reports/chart-section.js
