+++ js/src/pages/reports/compare-table-card.js
