+++ js/src/pages/reports/metric-number.js
