+++ js/src/pages/reports/metric-number.scss
