+++ js/src/pages/reports/products/compare-products-table-card.js
