+++ js/src/pages/reports/products/filterConfig.js
