+++ js/src/pages/reports/summary-section.js
